revert: Revert decompilation of func_8001BB4C
Reverted the C implementation of the complex jump-table function to restore the build to a stable state. This function will be revisited later. We're SO close to the end of this module, but I'm stuck in a loop because 8001BB4C keeps getting referenced by stuff we aren't accessing atm. Reverting temporarily to asm and returning once we detangle the rest of the functions mgiht yield better results.